add custom hook, add common state for todos, edit some naming
